Attempt to fix the nacl bots.
We're getting errors on certain native client bots where they error
during gclient sync while pulling down the dart sdk. This attempts
to fix it by checking to see if a dart sdk directory exists before
we remove it.

BUG=459376
R=ncbray@chromium.org

Review URL: https://codereview.chromium.org/943843003

+++ tools/dart/update.py
@@ -54,7 +54,8 @@ def main():
 
   if get_sdk:
     # Completely remove all traces of the previous SDK.
-    shutil.rmtree(DART_SDK_DIR)
+    if os.path.exists(DART_SDK_DIR):
+      shutil.rmtree(DART_SDK_DIR)
     os.mkdir(DART_SDK_DIR)
 
     wget_command = ['wget', '-N', '-c', LINUX_64_SDK, '-P', DART_SDK_DIR]
